feat: add notification sse network service tc

+++ src/main/java/com/example/BeGroom/notification/service/network/SseNotificationNetworkService.java
@@ -8,6 +8,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
 import java.io.IOException;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
 import java.util.List;
 import java.util.Map;
 
@@ -22,8 +24,13 @@ public class SseNotificationNetworkService implements NotificationNetworkService
     private Long defaultTimeout;
 
     @Override
-    public SseEmitter connect(Long memberId) {
-        String emitterId = memberId + "_" + System.currentTimeMillis();
+    public SseEmitter connect(Long memberId, LocalDateTime connectTime) {
+        if(memberId == null || memberId <= 0L){
+            throw new IllegalArgumentException("멤버 ID는 양수여야 합니다.");
+        }
+
+        long timestamp = connectTime.atZone(ZoneId.systemDefault()).toInstant().toEpochMilli();
+        String emitterId = memberId + "_" + timestamp;
         SseEmitter emitter = new SseEmitter(defaultTimeout);
 
         emitter.onCompletion(() -> {
@@ -52,19 +59,18 @@ public class SseNotificationNetworkService implements NotificationNetworkService
 
     @Override
     public void disconnect(Long memberId) {
-        // 1. 해당 회원의 모든 Emitter를 찾습니다.
+        if(memberId == null || memberId <= 0L){
+            throw new IllegalArgumentException("멤버 ID는 양수여야 합니다.");
+        }
+
         Map<String, SseEmitter> emitters = emitterRepository.findAllStartWithById(String.valueOf(memberId));
 
-        // 2. 루프를 돌면서 정리합니다.
         emitters.forEach((id, emitter) -> {
             try {
-                // [중요] 클라이언트에게 "연결 끝났다"고 기술적으로 끊어줍니다.
                 emitter.complete();
             } catch (Exception e) {
-                // 이미 끊긴 경우 등 에러 무시 (로그만 남김)
                 log.warn("disconnect fail : " + id);
             } finally {
-                // [중요] 저장소에서도 지워줍니다.
                 emitterRepository.deleteById(id);
             }
         });
@@ -96,9 +102,6 @@ public class SseNotificationNetworkService implements NotificationNetworkService
                     .data(data));
         } catch (IOException | IllegalStateException e) {
             emitterRepository.deleteById(id);
-            System.out.print("SSE 연결이 끊겨서 전송 실패. Emitter 삭제함: userId={}");
-            System.out.println(id);
-
         }
     }
